Colocate popover-specific component
