+++ assets/js/dashboard/keybinding.tsx
@@ -154,76 +154,3 @@ export function KeybindHint({
     </kbd>
   )
 }
-
-/**
- * Rendering this component captures the Escape key on targetRef.current,
- * blurring the element on Escape, and stopping the event from propagating.
- * Needed to prevent other Escape handlers that may exist from running.
- */
-export function BlurMenuButtonOnEscape({
-  targetRef,
-  ...props
-}: {
-  buttonId?: string
-  targetRef: RefObject<HTMLElement>
-}) {
-  const { registerDropmenuState } = useRoutelessModalsContext()
-
-  useEffect(() => {
-    const buttonId =
-      props.buttonId ?? `button-${Math.floor(Math.random() * 10000)}`
-
-    const observer = new MutationObserver((mutations) => {
-      mutations.forEach((mutation) => {
-        if (
-          mutation.type === 'attributes' &&
-          mutation.attributeName === 'data-open'
-        ) {
-          const element = mutation.target as Element
-          registerDropmenuState({
-            id: buttonId,
-            isOpen: element.hasAttribute('data-open')
-          })
-        }
-      })
-    })
-
-    const element = targetRef.current
-
-    if (element) {
-      registerDropmenuState({
-        id: buttonId,
-        isOpen: element.hasAttribute('data-open')
-      })
-      observer.observe(element, {
-        attributes: true,
-        attributeFilter: ['data-open']
-      })
-    }
-
-    return () => {
-      if (element) {
-        registerDropmenuState({ id: buttonId, isOpen: false })
-      }
-      observer.disconnect()
-    }
-  }, [targetRef, registerDropmenuState, props.buttonId])
-
-  return (
-    <Keybind
-      keyboardKey="Escape"
-      type="keyup"
-      handler={(event) => {
-        const t = event.target as HTMLElement | null
-        if (typeof t?.blur === 'function') {
-          if (t === targetRef.current) {
-            t.blur()
-            event.stopPropagation()
-          }
-        }
-      }}
-      targetRef={targetRef}
-      shouldIgnoreWhen={[isModifierPressed, isTyping]}
-    />
-  )
-}
