+++ assets/js/dashboard/nav-menu/query-periods/query-period-menu.tsx
@@ -3,7 +3,6 @@ import classNames from 'classnames'
 import { useQueryContext } from '../../query-context'
 import { useSiteContext } from '../../site-context'
 import {
-  BlurMenuButtonOnEscape,
   isModifierPressed,
   isTyping,
   Keybind,
@@ -25,7 +24,7 @@ import {
 import { useMatch } from 'react-router-dom'
 import { rootRoute } from '../../router'
 import { PopoverButton, PopoverPanel, Transition } from '@headlessui/react'
-import { popover } from '../../components/popover'
+import { popover, BlurMenuButtonOnEscape } from '../../components/popover'
 import {
   datemenuButtonClassName,
   DateMenuChevron,
